random size maze generation

+++ Character.h
@@ -4,7 +4,7 @@
 extern void initCharacterPosition(Coordinates);
 extern void drawCharacter();
 extern void moveCharacter(int, Maze);
-extern void setCharacterSize(int userSize);
+extern void setCharacterSize(float userSize);
 extern Coordinates getCharacterPosition();
 
 #endif
